requirements.txt
python-dotenv
langchain-core
langchain-community
langchain-openai
# langchain-anthropic
chromadb
streamlit
openai
pydantic
uuid
# pysqlite3-binary
langfuse
pypdf
datasets
ragas